Prevents multiple plans from being created with the same id
In the past, multiple calls to GetPlan could happen in parallel. `Items`
would not have been set yet for this plan but another call to the function
GetPlanCreateOptions might have been faster. This means that the second
call would return the same option leading to an error of the form "this
plan already exists"

We now take a lock when calling GetPlan to prevent this.