Adds the `FailureReason` property to `StripeRedirect`
We're currently missing the `FailureReason` property which is documented
here [1]. This patch adds it.

We do something slightly unusual (so far) here by adding reasons as a
new new called `StripeRedirectFailureReason`. There is only one other
documented instance of this in the library in `StripeBilling`, but we
might want to further pursue this pattern elsewhere in the future.

Fixes #1118.

[1] https://stripe.com/docs/api#source_object-redirect-failure_reason