Add support for multi-level Expand feature
All options class now have the ability to pass a list of properties to
expand. Until then, we had to add the ExpandXXXX property to the service
and it only supported one level.

One downside is that this could cause crashes if you're on a version of
the library that does not expand a specific property to be expandable.