Only run Coveralls if $COVERALLS_REPO_TOKEN is available
I noticed that builds from third party contributors apparently fail
because no Coveralls key is available (`$COVERALLS_REPO_TOKEN` is a
secure token and AppVeyor won't put it in all build environments). Here
we still run coverage, but only invoke the Coveralls step if a key is
available for it.